requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
Jinja2==3.1.6
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
SQLAlchemy==2.0.48

discord.py==2.7.1
groq==1.1.1
requests==2.32.5

google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
